cli/geolocation: move modules into doublezero-geolocation-cli crate
