+++ client/doublezero/src/cli/geolocation/mod.rs
@@ -1,21 +1,4 @@
-use clap::{Args, Subcommand};
+pub use doublezero_geolocation_cli::cli::GeolocationCliCommand;
 
 pub mod probe;
 pub mod user;
-
-use probe::ProbeCliCommand;
-use user::UserCliCommand;
-
-#[derive(Args, Debug)]
-pub struct GeolocationCliCommand {
-    #[command(subcommand)]
-    pub command: GeolocationCommands,
-}
-
-#[derive(Subcommand, Debug)]
-pub enum GeolocationCommands {
-    /// Manage geolocation probes
-    Probe(ProbeCliCommand),
-    /// Manage geolocation users and targets
-    User(UserCliCommand),
-}
