+++ crates/doublezero-geolocation-cli/src/client.rs
