+++ crates/doublezero-geolocation-cli/src/lib.rs
@@ -1,3 +1,12 @@
 //! RFC-20 module crate for the `doublezero geolocation` subcommand tree.
 //!
 //! See `rfcs/rfc20-cli-standardization.md` and `docs/cli-standard.md`.
+
+pub mod cli;
+pub mod client;
+pub mod init;
+pub mod probe;
+pub mod user;
+
+pub use cli::{GeolocationCliCommand, GeolocationCommand};
+pub use client::{GeoCliCommand, GeoCliCommandImpl};
