+++ crates/doublezero-geolocation-cli/src/probe/get.rs
@@ -1,4 +1,5 @@
-use crate::{geoclicommand::GeoCliCommand, validators::validate_pubkey_or_code};
+use crate::client::GeoCliCommand;
+use doublezero_cli_core::validators::validate_pubkey_or_code;
 use clap::Args;
 use doublezero_program_common::serializer;
 use doublezero_sdk::geolocation::geo_probe::get::GetGeoProbeCommand;
@@ -101,7 +102,7 @@ impl GetGeoProbeCliCommand {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::geoclicommand::MockGeoCliCommand;
+    use crate::client::MockGeoCliCommand;
     use doublezero_geolocation::state::{accounttype::AccountType, geo_probe::GeoProbe};
     use mockall::predicate;
     use solana_sdk::pubkey::Pubkey;
