+++ crates/doublezero-geolocation-cli/src/probe/update.rs
@@ -1,42 +1,57 @@
-use crate::{geoclicommand::GeoCliCommand, validators::validate_pubkey_or_code};
+use crate::client::GeoCliCommand;
+use doublezero_cli_core::validators::{validate_pubkey, validate_pubkey_or_code};
 use clap::Args;
 use doublezero_sdk::geolocation::geo_probe::{
-    delete::DeleteGeoProbeCommand, get::GetGeoProbeCommand,
+    get::GetGeoProbeCommand, update::UpdateGeoProbeCommand,
 };
-use std::io::Write;
+use solana_sdk::pubkey::Pubkey;
+use std::{io::Write, net::Ipv4Addr};
 
 #[derive(Args, Debug)]
-pub struct DeleteGeoProbeCliCommand {
-    /// Probe pubkey or code to delete
+pub struct UpdateGeoProbeCliCommand {
+    /// Probe pubkey or code to update
     #[arg(long, value_parser = validate_pubkey_or_code)]
     pub probe: String,
-    /// Skip confirmation prompt
-    #[arg(long, default_value_t = false)]
-    pub yes: bool,
+    /// Updated public IPv4 address
+    #[arg(long)]
+    pub public_ip: Option<Ipv4Addr>,
+    /// Updated UDP listen port for location offsets
+    #[arg(long)]
+    pub port: Option<u16>,
+    /// Updated signing public key
+    #[arg(long, value_parser = validate_pubkey)]
+    pub signing_pubkey: Option<String>,
 }
 
-impl DeleteGeoProbeCliCommand {
+impl UpdateGeoProbeCliCommand {
     pub fn execute<C: GeoCliCommand, W: Write>(self, client: &C, out: &mut W) -> eyre::Result<()> {
+        if self.public_ip.is_none() && self.port.is_none() && self.signing_pubkey.is_none() {
+            return Err(eyre::eyre!(
+                "At least one of --public-ip, --port, or --signing-pubkey is required"
+            ));
+        }
+
         let (_, resolved_probe) = client.get_geo_probe(GetGeoProbeCommand {
             pubkey_or_code: self.probe,
         })?;
         let code = resolved_probe.code;
 
-        if !self.yes {
-            eprint!("Delete probe '{code}'? [y/N]: ");
-            let mut input = String::new();
-            std::io::stdin().read_line(&mut input)?;
-            if !input.trim().eq_ignore_ascii_case("y") {
-                writeln!(out, "Aborted.")?;
-                return Ok(());
-            }
-        }
+        let metrics_publisher_pk = self
+            .signing_pubkey
+            .map(|mp| {
+                mp.parse::<Pubkey>()
+                    .map_err(|_| eyre::eyre!("invalid signing pubkey: {mp}"))
+            })
+            .transpose()?;
 
         let serviceability_globalstate_pk = client.get_serviceability_globalstate_pk();
 
-        let sig = client.delete_geo_probe(DeleteGeoProbeCommand {
+        let sig = client.update_geo_probe(UpdateGeoProbeCommand {
             code,
             serviceability_globalstate_pk,
+            public_ip: self.public_ip,
+            location_offset_port: self.port,
+            metrics_publisher_pk,
         })?;
 
         writeln!(out, "Signature: {sig}")?;
@@ -48,14 +63,13 @@ impl DeleteGeoProbeCliCommand {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::geoclicommand::MockGeoCliCommand;
+    use crate::client::MockGeoCliCommand;
     use doublezero_geolocation::state::{accounttype::AccountType, geo_probe::GeoProbe};
     use mockall::predicate;
     use solana_sdk::{pubkey::Pubkey, signature::Signature};
-    use std::net::Ipv4Addr;
 
     #[test]
-    fn test_cli_geo_probe_delete() {
+    fn test_cli_geo_probe_update() {
         let mut client = MockGeoCliCommand::new();
 
         let svc_gs_pk = Pubkey::from_str_const("HQ2UUt18uJqKaQFJhgV9zaTdQxUZjNrsKFgoEDquBkcx");
@@ -94,17 +108,22 @@ mod tests {
             .returning(move || svc_gs_pk);
 
         client
-            .expect_delete_geo_probe()
-            .with(predicate::eq(DeleteGeoProbeCommand {
+            .expect_update_geo_probe()
+            .with(predicate::eq(UpdateGeoProbeCommand {
                 code: "ams-probe-01".to_string(),
                 serviceability_globalstate_pk: svc_gs_pk,
+                public_ip: Some(Ipv4Addr::new(192, 168, 1, 1)),
+                location_offset_port: None,
+                metrics_publisher_pk: None,
             }))
             .returning(move |_| Ok(signature));
 
         let mut output = Vec::new();
-        let res = DeleteGeoProbeCliCommand {
+        let res = UpdateGeoProbeCliCommand {
             probe: "ams-probe-01".to_string(),
-            yes: true,
+            public_ip: Some(Ipv4Addr::new(192, 168, 1, 1)),
+            port: None,
+            signing_pubkey: None,
         }
         .execute(&client, &mut output);
         assert!(res.is_ok());
