+++ crates/doublezero-geolocation-cli/src/user/add_target.rs
@@ -1,7 +1,5 @@
-use crate::{
-    geoclicommand::GeoCliCommand,
-    validators::{validate_pubkey, validate_pubkey_or_code},
-};
+use crate::client::GeoCliCommand;
+use doublezero_cli_core::validators::{validate_pubkey, validate_pubkey_or_code};
 use clap::{Args, ValueEnum};
 use doublezero_geolocation::state::geolocation_user::GeoLocationTargetType;
 use doublezero_sdk::geolocation::{
@@ -137,7 +135,7 @@ pub(super) fn resolve_probe<C: GeoCliCommand>(
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::geoclicommand::MockGeoCliCommand;
+    use crate::client::MockGeoCliCommand;
     use doublezero_geolocation::state::{
         accounttype::AccountType,
         geo_probe::GeoProbe,
