+++ crates/doublezero-geolocation-cli/src/user/remove_target.rs
@@ -1,7 +1,5 @@
-use crate::{
-    geoclicommand::GeoCliCommand,
-    validators::{validate_pubkey, validate_pubkey_or_code},
-};
+use crate::client::GeoCliCommand;
+use doublezero_cli_core::validators::{validate_pubkey, validate_pubkey_or_code};
 use clap::Args;
 use doublezero_geolocation::state::geolocation_user::GeoLocationTargetType;
 use doublezero_sdk::geolocation::geolocation_user::{
@@ -84,7 +82,7 @@ impl RemoveTargetCliCommand {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::geoclicommand::MockGeoCliCommand;
+    use crate::client::MockGeoCliCommand;
     use doublezero_geolocation::state::{
         accounttype::AccountType,
         geo_probe::GeoProbe,
