+++ crates/doublezero-geolocation-cli/src/user/set_result_destination.rs
@@ -1,4 +1,5 @@
-use crate::{geoclicommand::GeoCliCommand, validators::validate_code};
+use crate::client::GeoCliCommand;
+use doublezero_cli_core::validators::validate_code;
 use clap::Args;
 use doublezero_geolocation::validation::validate_public_ip;
 use doublezero_sdk::geolocation::geolocation_user::{
@@ -115,7 +116,7 @@ impl SetResultDestinationCliCommand {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::geoclicommand::MockGeoCliCommand;
+    use crate::client::MockGeoCliCommand;
     use doublezero_geolocation::state::{
         accounttype::AccountType,
         geolocation_user::{
