+++ crates/doublezero-geolocation-cli/src/user/update.rs
@@ -1,4 +1,5 @@
-use crate::{geoclicommand::GeoCliCommand, validators::validate_pubkey_or_code};
+use crate::client::GeoCliCommand;
+use doublezero_cli_core::validators::validate_pubkey_or_code;
 use clap::Args;
 use doublezero_sdk::geolocation::geolocation_user::{
     get::GetGeolocationUserCommand, update::UpdateGeolocationUserCommand,
@@ -36,7 +37,7 @@ impl UpdateGeolocationUserCliCommand {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::geoclicommand::MockGeoCliCommand;
+    use crate::client::MockGeoCliCommand;
     use doublezero_geolocation::state::{
         accounttype::AccountType,
         geolocation_user::{
