+++ smartcontract/cli/Cargo.toml
@@ -39,7 +39,6 @@ tracing.workspace = true
 # Dependencies from this workspace
 doublezero-cli-core.workspace = true
 doublezero-config.workspace = true
-doublezero-geolocation = { workspace = true, features = ["no-entrypoint"] }
 doublezero-program-common.workspace = true
 doublezero_sdk.workspace = true
 doublezero-serviceability = { workspace = true, features = ["serde"] }
