--- smartcontract/cli/src/geolocation/config/mod.rs
@@ -1,2 +1,0 @@
-pub mod get;
-pub mod set;
