--- smartcontract/cli/src/geolocation/mod.rs
@@ -1,4 +1,0 @@
-pub mod config;
-pub mod probe;
-pub mod programconfig;
-pub mod user;
