--- smartcontract/cli/src/geolocation/probe/mod.rs
@@ -1,7 +1,0 @@
-pub mod add_parent;
-pub mod create;
-pub mod delete;
-pub mod get;
-pub mod list;
-pub mod remove_parent;
-pub mod update;
