--- smartcontract/cli/src/geolocation/programconfig/mod.rs
@@ -1,1 +1,0 @@
-pub mod init;
