--- smartcontract/cli/src/geolocation/user/mod.rs
@@ -1,9 +1,0 @@
-pub mod add_target;
-pub mod create;
-pub mod delete;
-pub mod get;
-pub mod list;
-pub mod remove_target;
-pub mod set_result_destination;
-pub mod update;
-pub mod update_payment_status;
